# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/dialog/{id}/list": {
            "get": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dialog"
                ],
                "summary": "Returns messages.",
                "parameters": [
                    {
                        "type": "string",
                        "description": " ",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dialog.GetMessagesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/dialog/{id}/send": {
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dialog"
                ],
                "summary": "Sends a message.",
                "parameters": [
                    {
                        "type": "string",
                        "description": " ",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": " ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dialog.SendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Sign in using user ID.",
                "parameters": [
                    {
                        "description": " ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/account.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/account.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "/user/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Registers a new user.",
                "parameters": [
                    {
                        "description": " ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/account.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/account.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "account.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "user_id"
            ],
            "properties": {
                "password": {
                    "description": "Password to the account.",
                    "type": "string",
                    "example": "123456"
                },
                "user_id": {
                    "description": "ID of the post in UUIDv4 format.",
                    "type": "string",
                    "example": "bbeb7da8-6d75-4419-9d94-91ec52bc506c"
                }
            }
        },
        "account.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "Authorization token to the account (JWT).",
                    "type": "string"
                }
            }
        },
        "account.RegisterRequest": {
            "type": "object",
            "required": [
                "biography",
                "birthdate",
                "city",
                "first_name",
                "gender",
                "password",
                "second_name"
            ],
            "properties": {
                "biography": {
                    "description": "Biography of the user.",
                    "type": "string",
                    "example": "Software developer"
                },
                "birthdate": {
                    "description": "Birthday in the format \"1990-12-31\".",
                    "type": "string",
                    "example": "1990-01-01"
                },
                "city": {
                    "description": "City of the user.",
                    "type": "string",
                    "example": "New York"
                },
                "first_name": {
                    "description": "First name of the user.",
                    "type": "string",
                    "example": "John"
                },
                "gender": {
                    "description": "Gender of the user (\"Male\" or \"Female\").",
                    "allOf": [
                        {
                            "$ref": "#/definitions/common.Gender"
                        }
                    ],
                    "example": "Male"
                },
                "password": {
                    "description": "Password to the account.",
                    "type": "string",
                    "example": "123456"
                },
                "second_name": {
                    "description": "Second name of the user.",
                    "type": "string",
                    "example": "Doe"
                }
            }
        },
        "account.RegisterResponse": {
            "type": "object",
            "properties": {
                "user_id": {
                    "description": "ID of the user in UUIDv4 format.",
                    "type": "string",
                    "example": "bbeb7da8-6d75-4419-9d94-91ec52bc506c"
                }
            }
        },
        "common.Gender": {
            "type": "string",
            "enum": [
                "Male",
                "Female"
            ],
            "x-enum-varnames": [
                "GenderMale",
                "GenderFemale"
            ]
        },
        "dialog.GetMessagesResponse": {
            "type": "object",
            "properties": {
                "messages": {
                    "description": "List of messages.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dialog.GetMessagesResponseItem"
                    }
                }
            }
        },
        "dialog.GetMessagesResponseItem": {
            "type": "object",
            "properties": {
                "from": {
                    "description": "ID of the message sender in UUIDv4 format.",
                    "type": "string"
                },
                "text": {
                    "description": "Content of the message.",
                    "type": "string"
                },
                "to": {
                    "description": "ID of the message recipient in UUIDv4 format.",
                    "type": "string"
                }
            }
        },
        "dialog.SendMessageRequest": {
            "type": "object",
            "required": [
                "text"
            ],
            "properties": {
                "text": {
                    "description": "Content of the message.",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}